Refactor location string to id
