+++ src/core/marketplace/application/CreatePost/CreatePostCommand.ts
@@ -43,13 +43,13 @@ export default class CreatePostCommand extends Command {
     public readonly price: string;
 
     /**
-     * The location or region associated with the post.
+     * The location or region id associated with the post.
      *
      * @public
      * @readonly
      * @type {string}
      */
-    public readonly location: string;
+    public readonly locationId: string;
 
     /**
      * The unique identifier of the seller creating the post.
@@ -86,7 +86,7 @@ export default class CreatePostCommand extends Command {
      * @param {string} title - The title of the post.
      * @param {string} description - The description of the post.
      * @param {string} price - The price of the item being sold in the post.
-     * @param {string} location - The location or region associated with the post.
+     * @param {string} locationId - The location or region id associated with the post.
      * @param {string} sellerId - The unique identifier of the seller creating the post.
      * @param {string} category - The category or type of the item being posted.
      * @param {string} photoUrl - The URL of the photo associated with the post.
@@ -96,7 +96,7 @@ export default class CreatePostCommand extends Command {
         title: string,
         description: string,
         price: string,
-        location: string,
+        locationId: string,
         sellerId: string,
         category: string,
         photoUrl: string
@@ -106,7 +106,7 @@ export default class CreatePostCommand extends Command {
         this.title = title;
         this.description = description;
         this.price = price;
-        this.location = location;
+        this.locationId = locationId;
         this.sellerId = sellerId;
         this.category = category;
         this.photoUrl = photoUrl;
