+++ src/core/marketplace/infrastructure/persistence/SBPostReadModelStore.ts
@@ -30,7 +30,7 @@ export default class SBPostReadModel extends SupabaseClient implements PostReadM
             description: post.description,
             category: post.category,
             price: post.price,
-            location: post.location,
+            location_id: post.locationId,
             user_id: post.sellerId,
             photo_url: post.photoUrl,
             is_moderated: post.isModerated,
@@ -67,7 +67,7 @@ export default class SBPostReadModel extends SupabaseClient implements PostReadM
                 data.title,
                 data.description,
                 data.price,
-                data.location,
+                data.location_id,
                 data.user_id,
                 data.category,
                 data.photoUrl,
@@ -93,7 +93,7 @@ export default class SBPostReadModel extends SupabaseClient implements PostReadM
             title: post.title,
             description: post.description,
             price: post.price,
-            location: post.location,
+            location_id: post.locationId,
             user_id: post.sellerId,
             category: post.category,
             photo_url: post.photoUrl,
